refactor(n-api): 라우팅 정책 초기화 로직 묶음
